requirements.txt
selenium
beautifulsoup4~=4.9.1
pandas~=1.3.1
PyYAML~=5.3.1
lxml
webdriver-manager
